requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.4.0=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.2.0=pypi_0
bzip2=1.0.8=h620ffc9_4
ca-certificates=2022.4.26=hca03da5_0
cachetools=5.3.0=pypi_0
certifi=2022.6.15=py310hca03da5_0
charset-normalizer=3.1.0=pypi_0
contourpy=1.0.7=pypi_0
cycler=0.11.0=pypi_0
data-utils=0.0.0=pypi_0
filelock=3.10.0=pypi_0
fonttools=4.39.2=pypi_0
frozenlist=1.3.3=pypi_0
fsspec=2023.3.0=pypi_0
future=0.18.3=pypi_0
google-auth=2.16.2=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.51.3=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.1.0=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
kaleido=0.2.1=pypi_0
kiwisolver=1.4.4=pypi_0
libcxx=12.0.0=hf6beb65_1
libffi=3.4.2=hc377ac9_4
llvmlite=0.39.1=pypi_0
markdown=3.4.1=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.1=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
ncurses=6.3=h1a28f6b_2
networkx=3.0=pypi_0
numba=0.56.4=pypi_0
numpy=1.23.5=pypi_0
oauthlib=3.2.2=pypi_0
openssl=1.1.1o=h1a28f6b_0
packaging=23.0=pypi_0
pandas=1.5.3=pypi_0
pillow=9.4.0=pypi_0
pip=21.2.4=py310hca03da5_0
plotly=5.13.1=pypi_0
position-vae-project=0.0.1=pypi_0
protobuf=4.22.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pydeprecate=0.3.1=pypi_0
pynndescent=0.5.8=pypi_0
pyparsing=3.0.9=pypi_0
python=3.10.4=hbdb9e5c_0
python-dateutil=2.8.2=pypi_0
pytorch-lightning=1.5.6=pypi_0
pytz=2023.2=pypi_0
pyyaml=6.0=pypi_0
readline=8.1.2=h1a28f6b_1
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
setup-utils=0.0.0=pypi_0
setuptools=61.2.0=py310hca03da5_0
six=1.16.0=pypi_0
sqlite=3.38.5=h1058600_0
sympy=1.11.1=pypi_0
tenacity=8.2.2=pypi_0
tensorboard=2.12.0=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=hb8d0fd4_0
torch=2.0.0=pypi_0
torchmetrics=0.11.4=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.15.1=pypi_0
tqdm=4.65.0=pypi_0
typing-extensions=4.5.0=pypi_0
tzdata=2022a=hda174b7_0
umap-learn=0.5.3=pypi_0
urllib3=1.26.15=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h1a28f6b_1
yaml=0.2.5=h1a28f6b_0
yarl=1.8.2=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.12=h5a0b063_2
